Rename files to start with plain letter
